% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{addCategoryBlocks}
\alias{addCategoryBlocks}
\title{Add list of category items}
\usage{
addCategoryBlocks(orient, input_id, color_map, values)
}
\arguments{
\item{orient}{Orientation of the legend. Can be \code{"bottom"} (default, horizontal with labels below), \code{"top"} (horizontal with labels above), \code{"left"} (vertical with labels on the left)
and \code{"right"} (vertical with labels on the right).}

\item{input_id}{The CSS class used to trigger interactions}

\item{color_map}{A list of colors from where the corresponding item color will be retrieved}

\item{values}{A list of values from which the corresponding item label will be retrieved}
}
\value{
A list of the same length as values, containing either \code{"span"} or \code{"div"} elements depending on the chosen orientation.
}
\description{
Add list of category items
}
\seealso{
\code{\link[=categoricalLegend]{categoricalLegend()}}
}
