% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{addColorStrips}
\alias{addColorStrips}
\title{Add list of colored strips}
\usage{
addColorStrips(n_strips, color_map, orient, pos_function, size, thickness = 20)
}
\arguments{
\item{n_strips}{Number of strips to be added}

\item{color_map}{A list of colors corresponding to the number of strips}

\item{orient}{Orientation of the legend. Can be \code{"bottom"} (default, horizontal with labels below), \code{"top"} (horizontal with labels above), \code{"left"} (vertical with labels on the left)
and \code{"right"} (vertical with labels on the right).}

\item{pos_function}{A function to convert from index number to pixels}

\item{size}{The length of the list in pixels}

\item{thickness}{The height or width of the list in pixels}
}
\value{
A list of SVG \code{rect} shapes.
}
\description{
Add list of colored strips
}
\seealso{
\code{\link[=numericLegend]{numericLegend()}}
}
