% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{categoryBlock}
\alias{categoryBlock}
\title{Add a color-label block}
\usage{
categoryBlock(i, values, tag_name, class, color_map)
}
\arguments{
\item{i}{The index of the item to be created}

\item{values}{A list of values from which the corresponding item label will be retrieved}

\item{tag_name}{An HTML element tag}

\item{class}{The HTML element class that will enable interaction}

\item{color_map}{A list of colors from where the corresponding item color will be retrieved}
}
\value{
An HTML element with pointer cursor, a colored square and a label
}
\description{
Add a color-label block
}
