% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{colorStrip}
\alias{colorStrip}
\title{Add color strip}
\usage{
colorStrip(color, x = 0, y = 0, width = 1, height = 30)
}
\arguments{
\item{color}{A valid CSS color name}

\item{x}{The x position of the \code{rect} shape relative to a container}

\item{y}{The y position of the \code{rect} shape relative to a container}

\item{width}{The width of the \code{rect}}

\item{height}{The height of the \code{rect}}
}
\value{
A \code{rect} element with the \code{color} argument as fill and stroke
}
\description{
Add color strip
}
