% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-input.R
\name{updateCategoricalFilter}
\alias{updateCategoricalFilter}
\title{Change a categorical legend in the client}
\usage{
updateCategoricalFilter(
  session,
  inputId,
  label = NULL,
  select = NULL,
  deselect = NULL
)
}
\arguments{
\item{session}{The \code{session} object passed to function given to
\code{shinyServer}.}

\item{inputId}{The id of the input object.}

\item{label}{The label to set for the input object.}

\item{select}{Items to be selected.}

\item{deselect}{Items to be deselected.}
}
\description{
Change a categorical legend in the client
}
\details{
This function only affects the label and the selection. Re-creating the items requires deleting and re-creating the legend using \link[shinyjs]{shinyjs}, for example.
}
\seealso{
\code{\link[=categoricalColorFilter]{categoricalColorFilter()}}
}
